+++ serde-shape/src/impls/mod.rs
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 mod container;
-#[cfg(feature = "std")]
 mod net;
 mod primitive;
 mod result;